+++ src/periphery/blue-public-allocator/interfaces/IBluePublicAllocator.sol
@@ -18,7 +18,7 @@ interface IBluePublicAllocator {
     event SetCanPullFromMarket(address indexed sender, address indexed vault, address adapter, MarketParams marketParams, bool canPullFromMarket);
     event SetCanPullFromIdle(address indexed sender, address indexed vault, bool canPullFromIdle);
     event SetIsActiveAdapter(address indexed sender, address indexed vault, address indexed adapter, bool isActiveAdapter);
-    event SetPenalty(address indexed sender, address indexed vault, uint256 newPenalty);
+    event SetPenalty(address indexed sender, address indexed vault, uint64 newPenalty);
     event Reallocate(address sender, address indexed vault, address deallocateAdapter, bytes32 indexed deallocateId, address allocateAdapter, bytes32 indexed allocateId, uint128 assets, uint256 penaltyAssets);
     event AllocateFromIdle(address indexed sender, address indexed vault, address adapter, bytes32 indexed allocateId, uint128 assets, uint256 penaltyAssets);
 
@@ -45,7 +45,7 @@ interface IBluePublicAllocator {
     function setAbsoluteCap(address vault, address adapter, MarketParams calldata marketParams, uint256 newAbsoluteCap) external;
     function setCanPullFromMarket(address vault, address adapter, MarketParams calldata marketParams, bool newCanPullFromMarket) external;
     function setCanPullFromIdle(address vault, bool newCanPullFromIdle) external;
-    function setPenalty(address vault, uint256 newPenalty) external;
+    function setPenalty(address vault, uint64 newPenalty) external;
     function reallocate(address vault, address deallocateAdapter, MarketParams calldata deallocateMarketParams, address allocateAdapter, MarketParams calldata allocateMarketParams, uint128 assets) external;
     function allocateFromIdle(address vault, address adapter, MarketParams calldata marketParams, uint128 assets) external;
 }
